% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_afficher_boxplot1.R
\name{afficher_boxplot1}
\alias{afficher_boxplot1}
\title{afficher_boxplot1}
\usage{
afficher_boxplot1(df, var)
}
\arguments{
\item{df}{un dataframe}

\item{var}{un character}
}
\value{
a graph ggplot
}
\description{
Une fonction qui affiche une box plot
}
\examples{
afficher_boxplot1(mtcars,"mpg")
}
