% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_afficher_courbe_lorenz.R
\name{afficher_courbe_lorenz}
\alias{afficher_courbe_lorenz}
\title{afficher_courbe_lorenz}
\usage{
afficher_courbe_lorenz(
  df,
  varnum,
  col = "darkred",
  lwd = 4,
  xlab = "Part cumulée des individus (en \%)",
  ylab = paste("Part cumulée de", varnum, "(en \%)"),
  main = " "
)
}
\arguments{
\item{df}{un dataframe}

\item{varnum}{un character}

\item{col}{un character}

\item{lwd}{un numeric}

\item{xlab}{un character}

\item{ylab}{un character}

\item{main}{un character}
}
\value{
a graph de type plot
}
\description{
Une fonction qui affiche la courbe de Lorenz dans une appli Shiny
}
\examples{
afficher_courbe_lorenz(grandile,"PATRIMOINE")
}
