% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_afficher_courbe_lorenz2.R
\name{afficher_courbe_lorenz2}
\alias{afficher_courbe_lorenz2}
\title{afficher_courbe_lorenz2}
\usage{
afficher_courbe_lorenz2(
  varnum,
  varnumlib,
  col = "darkblue",
  lwd = 4,
  xlab = "Part cumulée des individus (en \%)",
  ylab = paste("Part cumulée de", varnumlib, "(en \%)"),
  main = " "
)
}
\arguments{
\item{varnum}{un vecteur numérique}

\item{varnumlib}{un character}

\item{col}{un character}

\item{lwd}{un numeric}

\item{xlab}{un character}

\item{ylab}{un character}

\item{main}{un character}
}
\value{
un graphique de type plot
}
\description{
Une fonction qui affiche la courbe de Lorenz dans une appli Shiny
}
\examples{
x <- modifier_repartition2(grandile,"PATRIMOINE",dimin = FALSE,tx=80)
afficher_courbe_lorenz2(x,"PATRIMOINE")
}
