% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_afficher_histo1.R
\name{afficher_histo1}
\alias{afficher_histo1}
\title{afficher_histo1}
\usage{
afficher_histo1(df, var)
}
\arguments{
\item{df}{un dataframe}

\item{var}{un character}
}
\value{
a graph ggplot
}
\description{
Une fonction qui affiche un histogramme
}
\examples{
afficher_histo1(mtcars,"mpg")
}
