% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_afficher_histo3.R
\name{afficher_histo3}
\alias{afficher_histo3}
\title{afficher_histo3}
\usage{
afficher_histo3(df, varnum, nbclass, type_decoup, breaknum)
}
\arguments{
\item{df}{un dataframe}

\item{varnum}{un character}

\item{nbclass}{une valeur numerique}

\item{type_decoup}{un character}

\item{breaknum}{un vecteur numérique}
}
\value{
un dataframe
}
\description{
Une fonction qui renvoie un histogramme à partir des paramètres
renseignés dans l'application
}
\examples{
df <- grandile
x <- "PATRIMOINE"
y <- 20
z <- "largeur_fixe"
a <- calculer_bornes(df,x,y,z)
afficher_histo3(df,x,y,z,a)
}
