% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ajouter_valeurs_extremes.R
\name{ajouter_valeurs_extremes}
\alias{ajouter_valeurs_extremes}
\title{ajouter_valeurs_extremes}
\usage{
ajouter_valeurs_extremes(df, num, sup = TRUE, nb)
}
\arguments{
\item{df}{un dataframe}

\item{num}{un character}

\item{sup}{un booléan}

\item{nb}{une valeur numérique}
}
\value{
un vecteur numérique
}
\description{
Une fonction qui ajout des valeurs extremes supérieures
ou inférieures à un vecteur numérique
}
\examples{
ajouter_valeurs_extremes(mtcars,"disp",nb=10)
}
