% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_calculer_bornes.R
\name{calculer_bornes}
\alias{calculer_bornes}
\title{calculer_bornes}
\usage{
calculer_bornes(df, num, nbclass, type_decoup)
}
\arguments{
\item{df}{un dataframe}

\item{num}{un character}

\item{nbclass}{une valeur numerique}

\item{type_decoup}{un character}
}
\value{
un vecteur numerique
}
\description{
Une fonction qui calcule les bornes des classes à partir d'un vecteur numérique
}
\examples{
df <- grandile
x <- "PATRIMOINE"
y <- 5
z <- "quantiles"
calculer_bornes(df,x,y,z)
}
