% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_calculer_indicateurs1.R
\name{calculer_indicateurs1}
\alias{calculer_indicateurs1}
\title{calculer_indicateurs1}
\usage{
calculer_indicateurs1(df, var)
}
\arguments{
\item{df}{un dataframe}

\item{var}{un character}
}
\value{
un vecteur numérique
}
\description{
Une fonction qui calcule des indicateurs de tendance centrale
et de dispersion
}
\examples{
calculer_indicateurs1(mtcars,"mpg")
}
