% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_calculer_indicateurs_redistrib.R
\name{calculer_indicateurs_redistrib}
\alias{calculer_indicateurs_redistrib}
\title{calculer_indicateurs_redistrib}
\usage{
calculer_indicateurs_redistrib(df, var)
}
\arguments{
\item{df}{un dataframe}

\item{var}{un character}
}
\value{
un vecteur de character
}
\description{
Une fonction qui affiche un vecteur de character
}
\examples{
calculer_indicateurs_redistrib(grandile,"REV_DISPONIBLE")
}
