% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_calculer_indicateurs_redistrib_vect.R
\name{calculer_indicateurs_redistrib_vect}
\alias{calculer_indicateurs_redistrib_vect}
\title{calculer_indicateurs_redistrib_vect}
\usage{
calculer_indicateurs_redistrib_vect(varnum)
}
\arguments{
\item{varnum}{un vecteur numeric}
}
\value{
un vecteur numeric
}
\description{
Une fonction qui calcule des indicateurs de concentration
}
\examples{
x <- modifier_repartition(grandile,"REV_DISPONIBLE",dimin = FALSE,tx=10)
calculer_indicateurs_redistrib_vect(x)
}
