% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_formater_indicateurs1.R
\name{formater_indicateurs1}
\alias{formater_indicateurs1}
\title{formater_indicateurs1}
\usage{
formater_indicateurs1(vec)
}
\arguments{
\item{vec}{un vecteur numérique}
}
\value{
a vecteur de character
}
\description{
Une fonction qui formate les éléments d'un vecteur numérique
}
\examples{
calculer_indicateurs1(mtcars,"mpg") \%>\% formater_indicateurs1()
}
