% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_grandile.R
\docType{data}
\name{grandile}
\alias{grandile}
\title{grandile}
\format{
grandile est un data.frame constitué de 5418 ménages (rows) et de 16 variables (col).
}
\source{
\url{https://cefil.org/} {Site du Cefil}
}
\usage{
grandile
}
\description{
Le fichier grandile a été constitué par des formateurs du Cefil. C'est une base
de données pédagogique utilisé dans beaucoup d'exercices tout au long de leur parcours.
}
\author{
Jean \email{jean.sebban@insee.fr}
}
\keyword{data}
