% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_modifier_repartition.R
\name{modifier_repartition}
\alias{modifier_repartition}
\title{modifier_repartition}
\usage{
modifier_repartition(df, slideinput, dimin = TRUE, tx)
}
\arguments{
\item{df}{un dataframe}

\item{slideinput}{un character}

\item{dimin}{un bolean}

\item{tx}{un numeric}
}
\value{
un vecteur de numeric
}
\description{
Une fonction qui modifie la concentration d'une numérique
}
\examples{
modifier_repartition(grandile,"REV_DISPONIBLE",dimin = FALSE,tx=99)
}
