% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_petitile.R
\docType{data}
\name{petitile}
\alias{petitile}
\title{petitile}
\format{
grandile est un data.frame constitué de 2468 ménages (rows) et de 16 variables (col).
}
\source{
\url{https://cefil.org/} {Site du Cefil}
}
\usage{
petitile
}
\description{
Le fichier petitile a été constitué par des formateurs du Cefil. C'est une base
de données pédagogique utilisé dans beaucoup d'exercices tout au long de leur parcours.
}
\author{
Jean \email{jean.sebban@insee.fr}
}
\keyword{data}
