% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_select_class_df.R
\name{select_class_df}
\alias{select_class_df}
\title{select_class_df}
\usage{
select_class_df(df, type)
}
\arguments{
\item{df}{un dataframe}

\item{type}{un character}
}
\value{
a vecteur de character
}
\description{
Une fonction qui renvoie un vecteur de caractères avec le type de
variable demandée
}
\examples{
select_class_df(grandile,"numeric")
}
