% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_tabuler_bornes.R
\name{tabuler_bornes}
\alias{tabuler_bornes}
\title{tabuler_bornes}
\usage{
tabuler_bornes(num, nbclass, type_decoup)
}
\arguments{
\item{num}{un vecteur numérique}

\item{nbclass}{une valeur numerique}

\item{type_decoup}{un character}
}
\value{
un dataframe
}
\description{
Une fonction qui renvoie un tableau avec les classes à partir du vecteur
numérique des classes
}
\examples{
df <- grandile
x <- "REV_DISPONIBLE"
y <- 5
z <- "quantiles"
a <- calculer_bornes(df,x,y,z)
b <- tabuler_bornes(a,y,z)
}
